test(file): 🚧 fixed get file test
